third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/concat.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] concat two float16 1D tensors of same shape along axis 0
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat three float16 1D tensors of different 1st dimension along axis 0
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat four float16 1D tensors of same 1st dimension along axis 0
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat four float16 1D tensors of different 1st dimension along axis 0
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat two float16 2D tensors of same shape along axis 0
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat two float16 2D tensors of same others dimensions except different 1st dimension along axis 0
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat four float16 2D tensors of same shape along axis 0
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat two float16 2D tensors of same others dimensions except different 2nd dimension along axis 1
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat three float16 2D tensors of same shape along axis 1
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat four float16 2D tensors of same others dimensions except different 2nd dimension along axis 1
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat two float16 3D tensors of same others dimensions except different 1st dimension along axis 0
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat four float16 3D tensors of same others dimensions except different 2nd dimension along axis 1
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat three float16 3D tensors of same shape along axis 2
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat two float16 4D tensors of same others dimensions except different 1st dimension along axis 0
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat three float16 4D tensors of same shape along axis 1
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat three float16 4D tensors of same others dimensions except different 3rd dimension along axis 2
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
[FAIL] concat four float16 4D tensors of same others dimensions except different 4th dimension along axis 3
  assert_less_than_equal: assert_array_approx_equals_ulp: test concat float16 actual -2 should be close enough to expected -0.39453125 by ULP distance: expected a number less than or equal to 0 but got 2480
Harness: the test ran to completion.
